Use ionicons api rather than hardcoding
