+++ build.gradle
@@ -29,6 +29,12 @@ dependencies {
     }
     implementation 'net.java.dev.textile-j:textile-j:2.2.864'
     implementation 'org.apache.commons:commons-text:1.10.0'
+    implementation('io.jenkins.plugins:ionicons-api:74.v93d5eb_813d5f') {
+        artifact {
+            name = "ionicons-api"
+            type = "jar" // Force Gradle to use the JAR artifact
+        }
+    }
 
     testImplementation 'junit:junit:4.13.2'
 }
