+++ src/main/java/org/kohsuke/stapler/idea/symbols/IoniconsApiSymbolFinder.java
@@ -1,7 +1,7 @@
 package org.kohsuke.stapler.idea.symbols;
 
-import com.intellij.icons.AllIcons;
 import com.intellij.openapi.project.Project;
+import io.jenkins.plugins.ionicons.Ionicons;
 import org.w3c.dom.Document;
 import org.xml.sax.SAXException;
 
@@ -12,11 +12,17 @@ import javax.xml.parsers.ParserConfigurationException;
 import javax.xml.xpath.XPath;
 import javax.xml.xpath.XPathConstants;
 import javax.xml.xpath.XPathFactory;
+import java.io.BufferedReader;
 import java.io.File;
 import java.io.IOException;
-import java.util.HashSet;
+import java.io.InputStream;
+import java.io.InputStreamReader;
+import java.nio.charset.StandardCharsets;
 import java.util.Set;
 import java.util.stream.Collectors;
+
+import static org.kohsuke.stapler.idea.icons.Icons.convertSymbol;
+import static org.kohsuke.stapler.idea.icons.Icons.convertSymbolToIcon;
 
 public class IoniconsApiSymbolFinder implements SymbolFinder {
 
@@ -32,1368 +38,8 @@ public class IoniconsApiSymbolFinder implements SymbolFinder {
             return Set.of();
         }
 
-        Set<String> symbols = new HashSet<>();
-        symbols.add("accessibility-outline");
-        symbols.add("accessibility-sharp");
-        symbols.add("accessibility");
-        symbols.add("add-circle-outline");
-        symbols.add("add-circle-sharp");
-        symbols.add("add-circle");
-        symbols.add("add-outline");
-        symbols.add("add-sharp");
-        symbols.add("add");
-        symbols.add("airplane-outline");
-        symbols.add("airplane-sharp");
-        symbols.add("airplane");
-        symbols.add("alarm-outline");
-        symbols.add("alarm-sharp");
-        symbols.add("alarm");
-        symbols.add("albums-outline");
-        symbols.add("albums-sharp");
-        symbols.add("albums");
-        symbols.add("alert-circle-outline");
-        symbols.add("alert-circle-sharp");
-        symbols.add("alert-circle");
-        symbols.add("alert-outline");
-        symbols.add("alert-sharp");
-        symbols.add("alert");
-        symbols.add("american-football-outline");
-        symbols.add("american-football-sharp");
-        symbols.add("american-football");
-        symbols.add("analytics-outline");
-        symbols.add("analytics-sharp");
-        symbols.add("analytics");
-        symbols.add("aperture-outline");
-        symbols.add("aperture-sharp");
-        symbols.add("aperture");
-        symbols.add("apps-outline");
-        symbols.add("apps-sharp");
-        symbols.add("apps");
-        symbols.add("archive-outline");
-        symbols.add("archive-sharp");
-        symbols.add("archive");
-        symbols.add("arrow-back-circle-outline");
-        symbols.add("arrow-back-circle-sharp");
-        symbols.add("arrow-back-circle");
-        symbols.add("arrow-back-outline");
-        symbols.add("arrow-back-sharp");
-        symbols.add("arrow-back");
-        symbols.add("arrow-down-circle-outline");
-        symbols.add("arrow-down-circle-sharp");
-        symbols.add("arrow-down-circle");
-        symbols.add("arrow-down-left-box-outline");
-        symbols.add("arrow-down-left-box-sharp");
-        symbols.add("arrow-down-left-box");
-        symbols.add("arrow-down-outline");
-        symbols.add("arrow-down-right-box-outline");
-        symbols.add("arrow-down-right-box-sharp");
-        symbols.add("arrow-down-right-box");
-        symbols.add("arrow-down-sharp");
-        symbols.add("arrow-down");
-        symbols.add("arrow-forward-circle-outline");
-        symbols.add("arrow-forward-circle-sharp");
-        symbols.add("arrow-forward-circle");
-        symbols.add("arrow-forward-outline");
-        symbols.add("arrow-forward-sharp");
-        symbols.add("arrow-forward");
-        symbols.add("arrow-redo-circle-outline");
-        symbols.add("arrow-redo-circle-sharp");
-        symbols.add("arrow-redo-circle");
-        symbols.add("arrow-redo-outline");
-        symbols.add("arrow-redo-sharp");
-        symbols.add("arrow-redo");
-        symbols.add("arrow-undo-circle-outline");
-        symbols.add("arrow-undo-circle-sharp");
-        symbols.add("arrow-undo-circle");
-        symbols.add("arrow-undo-outline");
-        symbols.add("arrow-undo-sharp");
-        symbols.add("arrow-undo");
-        symbols.add("arrow-up-circle-outline");
-        symbols.add("arrow-up-circle-sharp");
-        symbols.add("arrow-up-circle");
-        symbols.add("arrow-up-left-box-outline");
-        symbols.add("arrow-up-left-box-sharp");
-        symbols.add("arrow-up-left-box");
-        symbols.add("arrow-up-outline");
-        symbols.add("arrow-up-right-box-outline");
-        symbols.add("arrow-up-right-box-sharp");
-        symbols.add("arrow-up-right-box");
-        symbols.add("arrow-up-sharp");
-        symbols.add("arrow-up");
-        symbols.add("at-circle-outline");
-        symbols.add("at-circle-sharp");
-        symbols.add("at-circle");
-        symbols.add("at-outline");
-        symbols.add("at-sharp");
-        symbols.add("at");
-        symbols.add("attach-outline");
-        symbols.add("attach-sharp");
-        symbols.add("attach");
-        symbols.add("backspace-outline");
-        symbols.add("backspace-sharp");
-        symbols.add("backspace");
-        symbols.add("bag-add-outline");
-        symbols.add("bag-add-sharp");
-        symbols.add("bag-add");
-        symbols.add("bag-check-outline");
-        symbols.add("bag-check-sharp");
-        symbols.add("bag-check");
-        symbols.add("bag-handle-outline");
-        symbols.add("bag-handle-sharp");
-        symbols.add("bag-handle");
-        symbols.add("bag-outline");
-        symbols.add("bag-remove-outline");
-        symbols.add("bag-remove-sharp");
-        symbols.add("bag-remove");
-        symbols.add("bag-sharp");
-        symbols.add("bag");
-        symbols.add("balloon-outline");
-        symbols.add("balloon-sharp");
-        symbols.add("balloon");
-        symbols.add("ban-outline");
-        symbols.add("ban-sharp");
-        symbols.add("ban");
-        symbols.add("bandage-outline");
-        symbols.add("bandage-sharp");
-        symbols.add("bandage");
-        symbols.add("bar-chart-outline");
-        symbols.add("bar-chart-sharp");
-        symbols.add("bar-chart");
-        symbols.add("barbell-outline");
-        symbols.add("barbell-sharp");
-        symbols.add("barbell");
-        symbols.add("barcode-outline");
-        symbols.add("barcode-sharp");
-        symbols.add("barcode");
-        symbols.add("baseball-outline");
-        symbols.add("baseball-sharp");
-        symbols.add("baseball");
-        symbols.add("basket-outline");
-        symbols.add("basket-sharp");
-        symbols.add("basket");
-        symbols.add("basketball-outline");
-        symbols.add("basketball-sharp");
-        symbols.add("basketball");
-        symbols.add("battery-charging-outline");
-        symbols.add("battery-charging-sharp");
-        symbols.add("battery-charging");
-        symbols.add("battery-dead-outline");
-        symbols.add("battery-dead-sharp");
-        symbols.add("battery-dead");
-        symbols.add("battery-full-outline");
-        symbols.add("battery-full-sharp");
-        symbols.add("battery-full");
-        symbols.add("battery-half-outline");
-        symbols.add("battery-half-sharp");
-        symbols.add("battery-half");
-        symbols.add("beaker-outline");
-        symbols.add("beaker-sharp");
-        symbols.add("beaker");
-        symbols.add("bed-outline");
-        symbols.add("bed-sharp");
-        symbols.add("bed");
-        symbols.add("beer-outline");
-        symbols.add("beer-sharp");
-        symbols.add("beer");
-        symbols.add("bicycle-outline");
-        symbols.add("bicycle-sharp");
-        symbols.add("bicycle");
-        symbols.add("binoculars-outline");
-        symbols.add("binoculars-sharp");
-        symbols.add("binoculars");
-        symbols.add("bluetooth-outline");
-        symbols.add("bluetooth-sharp");
-        symbols.add("bluetooth");
-        symbols.add("boat-outline");
-        symbols.add("boat-sharp");
-        symbols.add("boat");
-        symbols.add("body-outline");
-        symbols.add("body-sharp");
-        symbols.add("body");
-        symbols.add("bonfire-outline");
-        symbols.add("bonfire-sharp");
-        symbols.add("bonfire");
-        symbols.add("book-outline");
-        symbols.add("book-sharp");
-        symbols.add("book");
-        symbols.add("bookmark-outline");
-        symbols.add("bookmark-sharp");
-        symbols.add("bookmark");
-        symbols.add("bookmarks-outline");
-        symbols.add("bookmarks-sharp");
-        symbols.add("bookmarks");
-        symbols.add("bowling-ball-outline");
-        symbols.add("bowling-ball-sharp");
-        symbols.add("bowling-ball");
-        symbols.add("briefcase-outline");
-        symbols.add("briefcase-sharp");
-        symbols.add("briefcase");
-        symbols.add("browsers-outline");
-        symbols.add("browsers-sharp");
-        symbols.add("browsers");
-        symbols.add("brush-outline");
-        symbols.add("brush-sharp");
-        symbols.add("brush");
-        symbols.add("bug-outline");
-        symbols.add("bug-sharp");
-        symbols.add("bug");
-        symbols.add("build-outline");
-        symbols.add("build-sharp");
-        symbols.add("build");
-        symbols.add("bulb-outline");
-        symbols.add("bulb-sharp");
-        symbols.add("bulb");
-        symbols.add("bus-outline");
-        symbols.add("bus-sharp");
-        symbols.add("bus");
-        symbols.add("business-outline");
-        symbols.add("business-sharp");
-        symbols.add("business");
-        symbols.add("cafe-outline");
-        symbols.add("cafe-sharp");
-        symbols.add("cafe");
-        symbols.add("calculator-outline");
-        symbols.add("calculator-sharp");
-        symbols.add("calculator");
-        symbols.add("calendar-clear-outline");
-        symbols.add("calendar-clear-sharp");
-        symbols.add("calendar-clear");
-        symbols.add("calendar-number-outline");
-        symbols.add("calendar-number-sharp");
-        symbols.add("calendar-number");
-        symbols.add("calendar-outline");
-        symbols.add("calendar-sharp");
-        symbols.add("calendar");
-        symbols.add("call-outline");
-        symbols.add("call-sharp");
-        symbols.add("call");
-        symbols.add("camera-outline");
-        symbols.add("camera-reverse-outline");
-        symbols.add("camera-reverse-sharp");
-        symbols.add("camera-reverse");
-        symbols.add("camera-sharp");
-        symbols.add("camera");
-        symbols.add("car-outline");
-        symbols.add("car-sharp");
-        symbols.add("car-sport-outline");
-        symbols.add("car-sport-sharp");
-        symbols.add("car-sport");
-        symbols.add("car");
-        symbols.add("card-outline");
-        symbols.add("card-sharp");
-        symbols.add("card");
-        symbols.add("caret-back-circle-outline");
-        symbols.add("caret-back-circle-sharp");
-        symbols.add("caret-back-circle");
-        symbols.add("caret-back-outline");
-        symbols.add("caret-back-sharp");
-        symbols.add("caret-back");
-        symbols.add("caret-down-circle-outline");
-        symbols.add("caret-down-circle-sharp");
-        symbols.add("caret-down-circle");
-        symbols.add("caret-down-outline");
-        symbols.add("caret-down-sharp");
-        symbols.add("caret-down");
-        symbols.add("caret-forward-circle-outline");
-        symbols.add("caret-forward-circle-sharp");
-        symbols.add("caret-forward-circle");
-        symbols.add("caret-forward-outline");
-        symbols.add("caret-forward-sharp");
-        symbols.add("caret-forward");
-        symbols.add("caret-up-circle-outline");
-        symbols.add("caret-up-circle-sharp");
-        symbols.add("caret-up-circle");
-        symbols.add("caret-up-outline");
-        symbols.add("caret-up-sharp");
-        symbols.add("caret-up");
-        symbols.add("cart-outline");
-        symbols.add("cart-sharp");
-        symbols.add("cart");
-        symbols.add("cash-outline");
-        symbols.add("cash-sharp");
-        symbols.add("cash");
-        symbols.add("cellular-outline");
-        symbols.add("cellular-sharp");
-        symbols.add("cellular");
-        symbols.add("chatbox-ellipses-outline");
-        symbols.add("chatbox-ellipses-sharp");
-        symbols.add("chatbox-ellipses");
-        symbols.add("chatbox-outline");
-        symbols.add("chatbox-sharp");
-        symbols.add("chatbox");
-        symbols.add("chatbubble-ellipses-outline");
-        symbols.add("chatbubble-ellipses-sharp");
-        symbols.add("chatbubble-ellipses");
-        symbols.add("chatbubble-outline");
-        symbols.add("chatbubble-sharp");
-        symbols.add("chatbubble");
-        symbols.add("chatbubbles-outline");
-        symbols.add("chatbubbles-sharp");
-        symbols.add("chatbubbles");
-        symbols.add("checkbox-outline");
-        symbols.add("checkbox-sharp");
-        symbols.add("checkbox");
-        symbols.add("checkmark-circle-outline");
-        symbols.add("checkmark-circle-sharp");
-        symbols.add("checkmark-circle");
-        symbols.add("checkmark-done-circle-outline");
-        symbols.add("checkmark-done-circle-sharp");
-        symbols.add("checkmark-done-circle");
-        symbols.add("checkmark-done-outline");
-        symbols.add("checkmark-done-sharp");
-        symbols.add("checkmark-done");
-        symbols.add("checkmark-outline");
-        symbols.add("checkmark-sharp");
-        symbols.add("checkmark");
-        symbols.add("chevron-back-circle-outline");
-        symbols.add("chevron-back-circle-sharp");
-        symbols.add("chevron-back-circle");
-        symbols.add("chevron-back-outline");
-        symbols.add("chevron-back-sharp");
-        symbols.add("chevron-back");
-        symbols.add("chevron-collapse-outline");
-        symbols.add("chevron-collapse-sharp");
-        symbols.add("chevron-collapse");
-        symbols.add("chevron-down-circle-outline");
-        symbols.add("chevron-down-circle-sharp");
-        symbols.add("chevron-down-circle");
-        symbols.add("chevron-down-outline");
-        symbols.add("chevron-down-sharp");
-        symbols.add("chevron-down");
-        symbols.add("chevron-expand-outline");
-        symbols.add("chevron-expand-sharp");
-        symbols.add("chevron-expand");
-        symbols.add("chevron-forward-circle-outline");
-        symbols.add("chevron-forward-circle-sharp");
-        symbols.add("chevron-forward-circle");
-        symbols.add("chevron-forward-outline");
-        symbols.add("chevron-forward-sharp");
-        symbols.add("chevron-forward");
-        symbols.add("chevron-up-circle-outline");
-        symbols.add("chevron-up-circle-sharp");
-        symbols.add("chevron-up-circle");
-        symbols.add("chevron-up-outline");
-        symbols.add("chevron-up-sharp");
-        symbols.add("chevron-up");
-        symbols.add("clipboard-outline");
-        symbols.add("clipboard-sharp");
-        symbols.add("clipboard");
-        symbols.add("close-circle-outline");
-        symbols.add("close-circle-sharp");
-        symbols.add("close-circle");
-        symbols.add("close-outline");
-        symbols.add("close-sharp");
-        symbols.add("close");
-        symbols.add("cloud-circle-outline");
-        symbols.add("cloud-circle-sharp");
-        symbols.add("cloud-circle");
-        symbols.add("cloud-done-outline");
-        symbols.add("cloud-done-sharp");
-        symbols.add("cloud-done");
-        symbols.add("cloud-download-outline");
-        symbols.add("cloud-download-sharp");
-        symbols.add("cloud-download");
-        symbols.add("cloud-offline-outline");
-        symbols.add("cloud-offline-sharp");
-        symbols.add("cloud-offline");
-        symbols.add("cloud-outline");
-        symbols.add("cloud-sharp");
-        symbols.add("cloud-upload-outline");
-        symbols.add("cloud-upload-sharp");
-        symbols.add("cloud-upload");
-        symbols.add("cloud");
-        symbols.add("cloudy-night-outline");
-        symbols.add("cloudy-night-sharp");
-        symbols.add("cloudy-night");
-        symbols.add("cloudy-outline");
-        symbols.add("cloudy-sharp");
-        symbols.add("cloudy");
-        symbols.add("code-download-outline");
-        symbols.add("code-download-sharp");
-        symbols.add("code-download");
-        symbols.add("code-outline");
-        symbols.add("code-sharp");
-        symbols.add("code-slash-outline");
-        symbols.add("code-slash-sharp");
-        symbols.add("code-slash");
-        symbols.add("code-working-outline");
-        symbols.add("code-working-sharp");
-        symbols.add("code-working");
-        symbols.add("code");
-        symbols.add("cog-outline");
-        symbols.add("cog-sharp");
-        symbols.add("cog");
-        symbols.add("color-fill-outline");
-        symbols.add("color-fill-sharp");
-        symbols.add("color-fill");
-        symbols.add("color-filter-outline");
-        symbols.add("color-filter-sharp");
-        symbols.add("color-filter");
-        symbols.add("color-palette-outline");
-        symbols.add("color-palette-sharp");
-        symbols.add("color-palette");
-        symbols.add("color-wand-outline");
-        symbols.add("color-wand-sharp");
-        symbols.add("color-wand");
-        symbols.add("compass-outline");
-        symbols.add("compass-sharp");
-        symbols.add("compass");
-        symbols.add("construct-outline");
-        symbols.add("construct-sharp");
-        symbols.add("construct");
-        symbols.add("contract-outline");
-        symbols.add("contract-sharp");
-        symbols.add("contract");
-        symbols.add("contrast-outline");
-        symbols.add("contrast-sharp");
-        symbols.add("contrast");
-        symbols.add("copy-outline");
-        symbols.add("copy-sharp");
-        symbols.add("copy");
-        symbols.add("create-outline");
-        symbols.add("create-sharp");
-        symbols.add("create");
-        symbols.add("crop-outline");
-        symbols.add("crop-sharp");
-        symbols.add("crop");
-        symbols.add("cube-outline");
-        symbols.add("cube-sharp");
-        symbols.add("cube");
-        symbols.add("cut-outline");
-        symbols.add("cut-sharp");
-        symbols.add("cut");
-        symbols.add("desktop-outline");
-        symbols.add("desktop-sharp");
-        symbols.add("desktop");
-        symbols.add("diamond-outline");
-        symbols.add("diamond-sharp");
-        symbols.add("diamond");
-        symbols.add("dice-outline");
-        symbols.add("dice-sharp");
-        symbols.add("dice");
-        symbols.add("disc-outline");
-        symbols.add("disc-sharp");
-        symbols.add("disc");
-        symbols.add("document-attach-outline");
-        symbols.add("document-attach-sharp");
-        symbols.add("document-attach");
-        symbols.add("document-lock-outline");
-        symbols.add("document-lock-sharp");
-        symbols.add("document-lock");
-        symbols.add("document-outline");
-        symbols.add("document-sharp");
-        symbols.add("document-text-outline");
-        symbols.add("document-text-sharp");
-        symbols.add("document-text");
-        symbols.add("document");
-        symbols.add("documents-outline");
-        symbols.add("documents-sharp");
-        symbols.add("documents");
-        symbols.add("download-outline");
-        symbols.add("download-sharp");
-        symbols.add("download");
-        symbols.add("duplicate-outline");
-        symbols.add("duplicate-sharp");
-        symbols.add("duplicate");
-        symbols.add("ear-outline");
-        symbols.add("ear-sharp");
-        symbols.add("ear");
-        symbols.add("earth-outline");
-        symbols.add("earth-sharp");
-        symbols.add("earth");
-        symbols.add("easel-outline");
-        symbols.add("easel-sharp");
-        symbols.add("easel");
-        symbols.add("egg-outline");
-        symbols.add("egg-sharp");
-        symbols.add("egg");
-        symbols.add("ellipse-outline");
-        symbols.add("ellipse-sharp");
-        symbols.add("ellipse");
-        symbols.add("ellipsis-horizontal-circle-outline");
-        symbols.add("ellipsis-horizontal-circle-sharp");
-        symbols.add("ellipsis-horizontal-circle");
-        symbols.add("ellipsis-horizontal-outline");
-        symbols.add("ellipsis-horizontal-sharp");
-        symbols.add("ellipsis-horizontal");
-        symbols.add("ellipsis-vertical-circle-outline");
-        symbols.add("ellipsis-vertical-circle-sharp");
-        symbols.add("ellipsis-vertical-circle");
-        symbols.add("ellipsis-vertical-outline");
-        symbols.add("ellipsis-vertical-sharp");
-        symbols.add("ellipsis-vertical");
-        symbols.add("enter-outline");
-        symbols.add("enter-sharp");
-        symbols.add("enter");
-        symbols.add("exit-outline");
-        symbols.add("exit-sharp");
-        symbols.add("exit");
-        symbols.add("expand-outline");
-        symbols.add("expand-sharp");
-        symbols.add("expand");
-        symbols.add("extension-puzzle-outline");
-        symbols.add("extension-puzzle-sharp");
-        symbols.add("extension-puzzle");
-        symbols.add("eye-off-outline");
-        symbols.add("eye-off-sharp");
-        symbols.add("eye-off");
-        symbols.add("eye-outline");
-        symbols.add("eye-sharp");
-        symbols.add("eye");
-        symbols.add("eyedrop-outline");
-        symbols.add("eyedrop-sharp");
-        symbols.add("eyedrop");
-        symbols.add("fast-food-outline");
-        symbols.add("fast-food-sharp");
-        symbols.add("fast-food");
-        symbols.add("female-outline");
-        symbols.add("female-sharp");
-        symbols.add("female");
-        symbols.add("file-tray-full-outline");
-        symbols.add("file-tray-full-sharp");
-        symbols.add("file-tray-full");
-        symbols.add("file-tray-outline");
-        symbols.add("file-tray-sharp");
-        symbols.add("file-tray-stacked-outline");
-        symbols.add("file-tray-stacked-sharp");
-        symbols.add("file-tray-stacked");
-        symbols.add("file-tray");
-        symbols.add("film-outline");
-        symbols.add("film-sharp");
-        symbols.add("film");
-        symbols.add("filter-circle-outline");
-        symbols.add("filter-circle-sharp");
-        symbols.add("filter-circle");
-        symbols.add("filter-outline");
-        symbols.add("filter-sharp");
-        symbols.add("filter");
-        symbols.add("finger-print-outline");
-        symbols.add("finger-print-sharp");
-        symbols.add("finger-print");
-        symbols.add("fish-outline");
-        symbols.add("fish-sharp");
-        symbols.add("fish");
-        symbols.add("fitness-outline");
-        symbols.add("fitness-sharp");
-        symbols.add("fitness");
-        symbols.add("flag-outline");
-        symbols.add("flag-sharp");
-        symbols.add("flag");
-        symbols.add("flame-outline");
-        symbols.add("flame-sharp");
-        symbols.add("flame");
-        symbols.add("flash-off-outline");
-        symbols.add("flash-off-sharp");
-        symbols.add("flash-off");
-        symbols.add("flash-outline");
-        symbols.add("flash-sharp");
-        symbols.add("flash");
-        symbols.add("flashlight-outline");
-        symbols.add("flashlight-sharp");
-        symbols.add("flashlight");
-        symbols.add("flask-outline");
-        symbols.add("flask-sharp");
-        symbols.add("flask");
-        symbols.add("flower-outline");
-        symbols.add("flower-sharp");
-        symbols.add("flower");
-        symbols.add("folder-open-outline");
-        symbols.add("folder-open-sharp");
-        symbols.add("folder-open");
-        symbols.add("folder-outline");
-        symbols.add("folder-sharp");
-        symbols.add("folder");
-        symbols.add("football-outline");
-        symbols.add("football-sharp");
-        symbols.add("football");
-        symbols.add("footsteps-outline");
-        symbols.add("footsteps-sharp");
-        symbols.add("footsteps");
-        symbols.add("funnel-outline");
-        symbols.add("funnel-sharp");
-        symbols.add("funnel");
-        symbols.add("game-controller-outline");
-        symbols.add("game-controller-sharp");
-        symbols.add("game-controller");
-        symbols.add("gift-outline");
-        symbols.add("gift-sharp");
-        symbols.add("gift");
-        symbols.add("git-branch-outline");
-        symbols.add("git-branch-sharp");
-        symbols.add("git-branch");
-        symbols.add("git-commit-outline");
-        symbols.add("git-commit-sharp");
-        symbols.add("git-commit");
-        symbols.add("git-compare-outline");
-        symbols.add("git-compare-sharp");
-        symbols.add("git-compare");
-        symbols.add("git-merge-outline");
-        symbols.add("git-merge-sharp");
-        symbols.add("git-merge");
-        symbols.add("git-network-outline");
-        symbols.add("git-network-sharp");
-        symbols.add("git-network");
-        symbols.add("git-pull-request-outline");
-        symbols.add("git-pull-request-sharp");
-        symbols.add("git-pull-request");
-        symbols.add("glasses-outline");
-        symbols.add("glasses-sharp");
-        symbols.add("glasses");
-        symbols.add("globe-outline");
-        symbols.add("globe-sharp");
-        symbols.add("globe");
-        symbols.add("golf-outline");
-        symbols.add("golf-sharp");
-        symbols.add("golf");
-        symbols.add("grid-outline");
-        symbols.add("grid-sharp");
-        symbols.add("grid");
-        symbols.add("hammer-outline");
-        symbols.add("hammer-sharp");
-        symbols.add("hammer");
-        symbols.add("hand-left-outline");
-        symbols.add("hand-left-sharp");
-        symbols.add("hand-left");
-        symbols.add("hand-right-outline");
-        symbols.add("hand-right-sharp");
-        symbols.add("hand-right");
-        symbols.add("happy-outline");
-        symbols.add("happy-sharp");
-        symbols.add("happy");
-        symbols.add("hardware-chip-outline");
-        symbols.add("hardware-chip-sharp");
-        symbols.add("hardware-chip");
-        symbols.add("headset-outline");
-        symbols.add("headset-sharp");
-        symbols.add("headset");
-        symbols.add("heart-circle-outline");
-        symbols.add("heart-circle-sharp");
-        symbols.add("heart-circle");
-        symbols.add("heart-dislike-circle-outline");
-        symbols.add("heart-dislike-circle-sharp");
-        symbols.add("heart-dislike-circle");
-        symbols.add("heart-dislike-outline");
-        symbols.add("heart-dislike-sharp");
-        symbols.add("heart-dislike");
-        symbols.add("heart-half-outline");
-        symbols.add("heart-half-sharp");
-        symbols.add("heart-half");
-        symbols.add("heart-outline");
-        symbols.add("heart-sharp");
-        symbols.add("heart");
-        symbols.add("help-buoy-outline");
-        symbols.add("help-buoy-sharp");
-        symbols.add("help-buoy");
-        symbols.add("help-circle-outline");
-        symbols.add("help-circle-sharp");
-        symbols.add("help-circle");
-        symbols.add("help-outline");
-        symbols.add("help-sharp");
-        symbols.add("help");
-        symbols.add("home-outline");
-        symbols.add("home-sharp");
-        symbols.add("home");
-        symbols.add("hourglass-outline");
-        symbols.add("hourglass-sharp");
-        symbols.add("hourglass");
-        symbols.add("ice-cream-outline");
-        symbols.add("ice-cream-sharp");
-        symbols.add("ice-cream");
-        symbols.add("id-card-outline");
-        symbols.add("id-card-sharp");
-        symbols.add("id-card");
-        symbols.add("image-outline");
-        symbols.add("image-sharp");
-        symbols.add("image");
-        symbols.add("images-outline");
-        symbols.add("images-sharp");
-        symbols.add("images");
-        symbols.add("infinite-outline");
-        symbols.add("infinite-sharp");
-        symbols.add("infinite");
-        symbols.add("information-circle-outline");
-        symbols.add("information-circle-sharp");
-        symbols.add("information-circle");
-        symbols.add("information-outline");
-        symbols.add("information-sharp");
-        symbols.add("information");
-        symbols.add("invert-mode-outline");
-        symbols.add("invert-mode-sharp");
-        symbols.add("invert-mode");
-        symbols.add("jenkins");
-        symbols.add("journal-outline");
-        symbols.add("journal-sharp");
-        symbols.add("journal");
-        symbols.add("key-outline");
-        symbols.add("key-sharp");
-        symbols.add("key");
-        symbols.add("keypad-outline");
-        symbols.add("keypad-sharp");
-        symbols.add("keypad");
-        symbols.add("language-outline");
-        symbols.add("language-sharp");
-        symbols.add("language");
-        symbols.add("laptop-outline");
-        symbols.add("laptop-sharp");
-        symbols.add("laptop");
-        symbols.add("layers-outline");
-        symbols.add("layers-sharp");
-        symbols.add("layers");
-        symbols.add("leaf-outline");
-        symbols.add("leaf-sharp");
-        symbols.add("leaf");
-        symbols.add("library-outline");
-        symbols.add("library-sharp");
-        symbols.add("library");
-        symbols.add("link-outline");
-        symbols.add("link-sharp");
-        symbols.add("link");
-        symbols.add("list-circle-outline");
-        symbols.add("list-circle-sharp");
-        symbols.add("list-circle");
-        symbols.add("list-outline");
-        symbols.add("list-sharp");
-        symbols.add("list");
-        symbols.add("locate-outline");
-        symbols.add("locate-sharp");
-        symbols.add("locate");
-        symbols.add("location-outline");
-        symbols.add("location-sharp");
-        symbols.add("location");
-        symbols.add("lock-closed-outline");
-        symbols.add("lock-closed-sharp");
-        symbols.add("lock-closed");
-        symbols.add("lock-open-outline");
-        symbols.add("lock-open-sharp");
-        symbols.add("lock-open");
-        symbols.add("log-in-outline");
-        symbols.add("log-in-sharp");
-        symbols.add("log-in");
-        symbols.add("log-out-outline");
-        symbols.add("log-out-sharp");
-        symbols.add("log-out");
-        symbols.add("logo-alipay");
-        symbols.add("logo-amazon");
-        symbols.add("logo-amplify");
-        symbols.add("logo-android");
-        symbols.add("logo-angular");
-        symbols.add("logo-appflow");
-        symbols.add("logo-apple-appstore");
-        symbols.add("logo-apple-ar");
-        symbols.add("logo-apple");
-        symbols.add("logo-behance");
-        symbols.add("logo-bitbucket");
-        symbols.add("logo-bitcoin");
-        symbols.add("logo-buffer");
-        symbols.add("logo-capacitor");
-        symbols.add("logo-chrome");
-        symbols.add("logo-closed-captioning");
-        symbols.add("logo-codepen");
-        symbols.add("logo-css3");
-        symbols.add("logo-designernews");
-        symbols.add("logo-deviantart");
-        symbols.add("logo-discord");
-        symbols.add("logo-docker");
-        symbols.add("logo-dribbble");
-        symbols.add("logo-dropbox");
-        symbols.add("logo-edge");
-        symbols.add("logo-electron");
-        symbols.add("logo-euro");
-        symbols.add("logo-facebook");
-        symbols.add("logo-figma");
-        symbols.add("logo-firebase");
-        symbols.add("logo-firefox");
-        symbols.add("logo-flickr");
-        symbols.add("logo-foursquare");
-        symbols.add("logo-github");
-        symbols.add("logo-gitlab");
-        symbols.add("logo-google-playstore");
-        symbols.add("logo-google");
-        symbols.add("logo-hackernews");
-        symbols.add("logo-html5");
-        symbols.add("logo-instagram");
-        symbols.add("logo-ionic");
-        symbols.add("logo-ionitron");
-        symbols.add("logo-javascript");
-        symbols.add("logo-laravel");
-        symbols.add("logo-linkedin");
-        symbols.add("logo-markdown");
-        symbols.add("logo-mastodon");
-        symbols.add("logo-medium");
-        symbols.add("logo-microsoft");
-        symbols.add("logo-no-smoking");
-        symbols.add("logo-nodejs");
-        symbols.add("logo-npm");
-        symbols.add("logo-octocat");
-        symbols.add("logo-paypal");
-        symbols.add("logo-pinterest");
-        symbols.add("logo-playstation");
-        symbols.add("logo-pwa");
-        symbols.add("logo-python");
-        symbols.add("logo-react");
-        symbols.add("logo-reddit");
-        symbols.add("logo-rss");
-        symbols.add("logo-sass");
-        symbols.add("logo-skype");
-        symbols.add("logo-slack");
-        symbols.add("logo-snapchat");
-        symbols.add("logo-soundcloud");
-        symbols.add("logo-stackoverflow");
-        symbols.add("logo-steam");
-        symbols.add("logo-stencil");
-        symbols.add("logo-tableau");
-        symbols.add("logo-tiktok");
-        symbols.add("logo-trapeze");
-        symbols.add("logo-tumblr");
-        symbols.add("logo-tux");
-        symbols.add("logo-twitch");
-        symbols.add("logo-twitter");
-        symbols.add("logo-usd");
-        symbols.add("logo-venmo");
-        symbols.add("logo-vercel");
-        symbols.add("logo-vimeo");
-        symbols.add("logo-vk");
-        symbols.add("logo-vue");
-        symbols.add("logo-web-component");
-        symbols.add("logo-wechat");
-        symbols.add("logo-whatsapp");
-        symbols.add("logo-windows");
-        symbols.add("logo-wordpress");
-        symbols.add("logo-x");
-        symbols.add("logo-xbox");
-        symbols.add("logo-xing");
-        symbols.add("logo-yahoo");
-        symbols.add("logo-yen");
-        symbols.add("logo-youtube");
-        symbols.add("magnet-outline");
-        symbols.add("magnet-sharp");
-        symbols.add("magnet");
-        symbols.add("mail-open-outline");
-        symbols.add("mail-open-sharp");
-        symbols.add("mail-open");
-        symbols.add("mail-outline");
-        symbols.add("mail-sharp");
-        symbols.add("mail-unread-outline");
-        symbols.add("mail-unread-sharp");
-        symbols.add("mail-unread");
-        symbols.add("mail");
-        symbols.add("male-female-outline");
-        symbols.add("male-female-sharp");
-        symbols.add("male-female");
-        symbols.add("male-outline");
-        symbols.add("male-sharp");
-        symbols.add("male");
-        symbols.add("man-outline");
-        symbols.add("man-sharp");
-        symbols.add("man");
-        symbols.add("map-outline");
-        symbols.add("map-sharp");
-        symbols.add("map");
-        symbols.add("medal-outline");
-        symbols.add("medal-sharp");
-        symbols.add("medal");
-        symbols.add("medical-outline");
-        symbols.add("medical-sharp");
-        symbols.add("medical");
-        symbols.add("medkit-outline");
-        symbols.add("medkit-sharp");
-        symbols.add("medkit");
-        symbols.add("megaphone-outline");
-        symbols.add("megaphone-sharp");
-        symbols.add("megaphone");
-        symbols.add("menu-outline");
-        symbols.add("menu-sharp");
-        symbols.add("menu");
-        symbols.add("mic-circle-outline");
-        symbols.add("mic-circle-sharp");
-        symbols.add("mic-circle");
-        symbols.add("mic-off-circle-outline");
-        symbols.add("mic-off-circle-sharp");
-        symbols.add("mic-off-circle");
-        symbols.add("mic-off-outline");
-        symbols.add("mic-off-sharp");
-        symbols.add("mic-off");
-        symbols.add("mic-outline");
-        symbols.add("mic-sharp");
-        symbols.add("mic");
-        symbols.add("moon-outline");
-        symbols.add("moon-sharp");
-        symbols.add("moon");
-        symbols.add("move-outline");
-        symbols.add("move-sharp");
-        symbols.add("move");
-        symbols.add("musical-note-outline");
-        symbols.add("musical-note-sharp");
-        symbols.add("musical-note");
-        symbols.add("musical-notes-outline");
-        symbols.add("musical-notes-sharp");
-        symbols.add("musical-notes");
-        symbols.add("navigate-circle-outline");
-        symbols.add("navigate-circle-sharp");
-        symbols.add("navigate-circle");
-        symbols.add("navigate-outline");
-        symbols.add("navigate-sharp");
-        symbols.add("navigate");
-        symbols.add("newspaper-outline");
-        symbols.add("newspaper-sharp");
-        symbols.add("newspaper");
-        symbols.add("notifications-circle-outline");
-        symbols.add("notifications-circle-sharp");
-        symbols.add("notifications-circle");
-        symbols.add("notifications-off-circle-outline");
-        symbols.add("notifications-off-circle-sharp");
-        symbols.add("notifications-off-circle");
-        symbols.add("notifications-off-outline");
-        symbols.add("notifications-off-sharp");
-        symbols.add("notifications-off");
-        symbols.add("notifications-outline");
-        symbols.add("notifications-sharp");
-        symbols.add("notifications");
-        symbols.add("nuclear-outline");
-        symbols.add("nuclear-sharp");
-        symbols.add("nuclear");
-        symbols.add("nutrition-outline");
-        symbols.add("nutrition-sharp");
-        symbols.add("nutrition");
-        symbols.add("open-outline");
-        symbols.add("open-sharp");
-        symbols.add("open");
-        symbols.add("options-outline");
-        symbols.add("options-sharp");
-        symbols.add("options");
-        symbols.add("paper-plane-outline");
-        symbols.add("paper-plane-sharp");
-        symbols.add("paper-plane");
-        symbols.add("partly-sunny-outline");
-        symbols.add("partly-sunny-sharp");
-        symbols.add("partly-sunny");
-        symbols.add("pause-circle-outline");
-        symbols.add("pause-circle-sharp");
-        symbols.add("pause-circle");
-        symbols.add("pause-outline");
-        symbols.add("pause-sharp");
-        symbols.add("pause");
-        symbols.add("paw-outline");
-        symbols.add("paw-sharp");
-        symbols.add("paw");
-        symbols.add("pencil-outline");
-        symbols.add("pencil-sharp");
-        symbols.add("pencil");
-        symbols.add("people-circle-outline");
-        symbols.add("people-circle-sharp");
-        symbols.add("people-circle");
-        symbols.add("people-outline");
-        symbols.add("people-sharp");
-        symbols.add("people");
-        symbols.add("person-add-outline");
-        symbols.add("person-add-sharp");
-        symbols.add("person-add");
-        symbols.add("person-circle-outline");
-        symbols.add("person-circle-sharp");
-        symbols.add("person-circle");
-        symbols.add("person-outline");
-        symbols.add("person-remove-outline");
-        symbols.add("person-remove-sharp");
-        symbols.add("person-remove");
-        symbols.add("person-sharp");
-        symbols.add("person");
-        symbols.add("phone-landscape-outline");
-        symbols.add("phone-landscape-sharp");
-        symbols.add("phone-landscape");
-        symbols.add("phone-portrait-outline");
-        symbols.add("phone-portrait-sharp");
-        symbols.add("phone-portrait");
-        symbols.add("pie-chart-outline");
-        symbols.add("pie-chart-sharp");
-        symbols.add("pie-chart");
-        symbols.add("pin-outline");
-        symbols.add("pin-sharp");
-        symbols.add("pin");
-        symbols.add("pint-outline");
-        symbols.add("pint-sharp");
-        symbols.add("pint");
-        symbols.add("pizza-outline");
-        symbols.add("pizza-sharp");
-        symbols.add("pizza");
-        symbols.add("planet-outline");
-        symbols.add("planet-sharp");
-        symbols.add("planet");
-        symbols.add("play-back-circle-outline");
-        symbols.add("play-back-circle-sharp");
-        symbols.add("play-back-circle");
-        symbols.add("play-back-outline");
-        symbols.add("play-back-sharp");
-        symbols.add("play-back");
-        symbols.add("play-circle-outline");
-        symbols.add("play-circle-sharp");
-        symbols.add("play-circle");
-        symbols.add("play-forward-circle-outline");
-        symbols.add("play-forward-circle-sharp");
-        symbols.add("play-forward-circle");
-        symbols.add("play-forward-outline");
-        symbols.add("play-forward-sharp");
-        symbols.add("play-forward");
-        symbols.add("play-outline");
-        symbols.add("play-sharp");
-        symbols.add("play-skip-back-circle-outline");
-        symbols.add("play-skip-back-circle-sharp");
-        symbols.add("play-skip-back-circle");
-        symbols.add("play-skip-back-outline");
-        symbols.add("play-skip-back-sharp");
-        symbols.add("play-skip-back");
-        symbols.add("play-skip-forward-circle-outline");
-        symbols.add("play-skip-forward-circle-sharp");
-        symbols.add("play-skip-forward-circle");
-        symbols.add("play-skip-forward-outline");
-        symbols.add("play-skip-forward-sharp");
-        symbols.add("play-skip-forward");
-        symbols.add("play");
-        symbols.add("podium-outline");
-        symbols.add("podium-sharp");
-        symbols.add("podium");
-        symbols.add("power-outline");
-        symbols.add("power-sharp");
-        symbols.add("power");
-        symbols.add("pricetag-outline");
-        symbols.add("pricetag-sharp");
-        symbols.add("pricetag");
-        symbols.add("pricetags-outline");
-        symbols.add("pricetags-sharp");
-        symbols.add("pricetags");
-        symbols.add("print-outline");
-        symbols.add("print-sharp");
-        symbols.add("print");
-        symbols.add("prism-outline");
-        symbols.add("prism-sharp");
-        symbols.add("prism");
-        symbols.add("pulse-outline");
-        symbols.add("pulse-sharp");
-        symbols.add("pulse");
-        symbols.add("push-outline");
-        symbols.add("push-sharp");
-        symbols.add("push");
-        symbols.add("qr-code-outline");
-        symbols.add("qr-code-sharp");
-        symbols.add("qr-code");
-        symbols.add("radio-button-off-outline");
-        symbols.add("radio-button-off-sharp");
-        symbols.add("radio-button-off");
-        symbols.add("radio-button-on-outline");
-        symbols.add("radio-button-on-sharp");
-        symbols.add("radio-button-on");
-        symbols.add("radio-outline");
-        symbols.add("radio-sharp");
-        symbols.add("radio");
-        symbols.add("rainy-outline");
-        symbols.add("rainy-sharp");
-        symbols.add("rainy");
-        symbols.add("reader-outline");
-        symbols.add("reader-sharp");
-        symbols.add("reader");
-        symbols.add("receipt-outline");
-        symbols.add("receipt-sharp");
-        symbols.add("receipt");
-        symbols.add("recording-outline");
-        symbols.add("recording-sharp");
-        symbols.add("recording");
-        symbols.add("refresh-circle-outline");
-        symbols.add("refresh-circle-sharp");
-        symbols.add("refresh-circle");
-        symbols.add("refresh-outline");
-        symbols.add("refresh-sharp");
-        symbols.add("refresh");
-        symbols.add("reload-circle-outline");
-        symbols.add("reload-circle-sharp");
-        symbols.add("reload-circle");
-        symbols.add("reload-outline");
-        symbols.add("reload-sharp");
-        symbols.add("reload");
-        symbols.add("remove-circle-outline");
-        symbols.add("remove-circle-sharp");
-        symbols.add("remove-circle");
-        symbols.add("remove-outline");
-        symbols.add("remove-sharp");
-        symbols.add("remove");
-        symbols.add("reorder-four-outline");
-        symbols.add("reorder-four-sharp");
-        symbols.add("reorder-four");
-        symbols.add("reorder-three-outline");
-        symbols.add("reorder-three-sharp");
-        symbols.add("reorder-three");
-        symbols.add("reorder-two-outline");
-        symbols.add("reorder-two-sharp");
-        symbols.add("reorder-two");
-        symbols.add("repeat-outline");
-        symbols.add("repeat-sharp");
-        symbols.add("repeat");
-        symbols.add("resize-outline");
-        symbols.add("resize-sharp");
-        symbols.add("resize");
-        symbols.add("restaurant-outline");
-        symbols.add("restaurant-sharp");
-        symbols.add("restaurant");
-        symbols.add("return-down-back-outline");
-        symbols.add("return-down-back-sharp");
-        symbols.add("return-down-back");
-        symbols.add("return-down-forward-outline");
-        symbols.add("return-down-forward-sharp");
-        symbols.add("return-down-forward");
-        symbols.add("return-up-back-outline");
-        symbols.add("return-up-back-sharp");
-        symbols.add("return-up-back");
-        symbols.add("return-up-forward-outline");
-        symbols.add("return-up-forward-sharp");
-        symbols.add("return-up-forward");
-        symbols.add("ribbon-outline");
-        symbols.add("ribbon-sharp");
-        symbols.add("ribbon");
-        symbols.add("rocket-outline");
-        symbols.add("rocket-sharp");
-        symbols.add("rocket");
-        symbols.add("rose-outline");
-        symbols.add("rose-sharp");
-        symbols.add("rose");
-        symbols.add("sad-outline");
-        symbols.add("sad-sharp");
-        symbols.add("sad");
-        symbols.add("save-outline");
-        symbols.add("save-sharp");
-        symbols.add("save");
-        symbols.add("scale-outline");
-        symbols.add("scale-sharp");
-        symbols.add("scale");
-        symbols.add("scan-circle-outline");
-        symbols.add("scan-circle-sharp");
-        symbols.add("scan-circle");
-        symbols.add("scan-outline");
-        symbols.add("scan-sharp");
-        symbols.add("scan");
-        symbols.add("school-outline");
-        symbols.add("school-sharp");
-        symbols.add("school");
-        symbols.add("search-circle-outline");
-        symbols.add("search-circle-sharp");
-        symbols.add("search-circle");
-        symbols.add("search-outline");
-        symbols.add("search-sharp");
-        symbols.add("search");
-        symbols.add("send-outline");
-        symbols.add("send-sharp");
-        symbols.add("send");
-        symbols.add("server-outline");
-        symbols.add("server-sharp");
-        symbols.add("server");
-        symbols.add("settings-outline");
-        symbols.add("settings-sharp");
-        symbols.add("settings");
-        symbols.add("shapes-outline");
-        symbols.add("shapes-sharp");
-        symbols.add("shapes");
-        symbols.add("share-outline");
-        symbols.add("share-sharp");
-        symbols.add("share-social-outline");
-        symbols.add("share-social-sharp");
-        symbols.add("share-social");
-        symbols.add("share");
-        symbols.add("shield-checkmark-outline");
-        symbols.add("shield-checkmark-sharp");
-        symbols.add("shield-checkmark");
-        symbols.add("shield-half-outline");
-        symbols.add("shield-half-sharp");
-        symbols.add("shield-half");
-        symbols.add("shield-outline");
-        symbols.add("shield-sharp");
-        symbols.add("shield");
-        symbols.add("shirt-outline");
-        symbols.add("shirt-sharp");
-        symbols.add("shirt");
-        symbols.add("shuffle-outline");
-        symbols.add("shuffle-sharp");
-        symbols.add("shuffle");
-        symbols.add("skull-outline");
-        symbols.add("skull-sharp");
-        symbols.add("skull");
-        symbols.add("snow-outline");
-        symbols.add("snow-sharp");
-        symbols.add("snow");
-        symbols.add("sparkles-outline");
-        symbols.add("sparkles-sharp");
-        symbols.add("sparkles");
-        symbols.add("speedometer-outline");
-        symbols.add("speedometer-sharp");
-        symbols.add("speedometer");
-        symbols.add("square-outline");
-        symbols.add("square-sharp");
-        symbols.add("square");
-        symbols.add("star-half-outline");
-        symbols.add("star-half-sharp");
-        symbols.add("star-half");
-        symbols.add("star-outline");
-        symbols.add("star-sharp");
-        symbols.add("star");
-        symbols.add("stats-chart-outline");
-        symbols.add("stats-chart-sharp");
-        symbols.add("stats-chart");
-        symbols.add("stop-circle-outline");
-        symbols.add("stop-circle-sharp");
-        symbols.add("stop-circle");
-        symbols.add("stop-outline");
-        symbols.add("stop-sharp");
-        symbols.add("stop");
-        symbols.add("stopwatch-outline");
-        symbols.add("stopwatch-sharp");
-        symbols.add("stopwatch");
-        symbols.add("storefront-outline");
-        symbols.add("storefront-sharp");
-        symbols.add("storefront");
-        symbols.add("subway-outline");
-        symbols.add("subway-sharp");
-        symbols.add("subway");
-        symbols.add("sunny-outline");
-        symbols.add("sunny-sharp");
-        symbols.add("sunny");
-        symbols.add("swap-horizontal-outline");
-        symbols.add("swap-horizontal-sharp");
-        symbols.add("swap-horizontal");
-        symbols.add("swap-vertical-outline");
-        symbols.add("swap-vertical-sharp");
-        symbols.add("swap-vertical");
-        symbols.add("sync-circle-outline");
-        symbols.add("sync-circle-sharp");
-        symbols.add("sync-circle");
-        symbols.add("sync-outline");
-        symbols.add("sync-sharp");
-        symbols.add("sync");
-        symbols.add("tablet-landscape-outline");
-        symbols.add("tablet-landscape-sharp");
-        symbols.add("tablet-landscape");
-        symbols.add("tablet-portrait-outline");
-        symbols.add("tablet-portrait-sharp");
-        symbols.add("tablet-portrait");
-        symbols.add("telescope-outline");
-        symbols.add("telescope-sharp");
-        symbols.add("telescope");
-        symbols.add("tennisball-outline");
-        symbols.add("tennisball-sharp");
-        symbols.add("tennisball");
-        symbols.add("terminal-outline");
-        symbols.add("terminal-sharp");
-        symbols.add("terminal");
-        symbols.add("text-outline");
-        symbols.add("text-sharp");
-        symbols.add("text");
-        symbols.add("thermometer-outline");
-        symbols.add("thermometer-sharp");
-        symbols.add("thermometer");
-        symbols.add("thumbs-down-outline");
-        symbols.add("thumbs-down-sharp");
-        symbols.add("thumbs-down");
-        symbols.add("thumbs-up-outline");
-        symbols.add("thumbs-up-sharp");
-        symbols.add("thumbs-up");
-        symbols.add("thunderstorm-outline");
-        symbols.add("thunderstorm-sharp");
-        symbols.add("thunderstorm");
-        symbols.add("ticket-outline");
-        symbols.add("ticket-sharp");
-        symbols.add("ticket");
-        symbols.add("time-outline");
-        symbols.add("time-sharp");
-        symbols.add("time");
-        symbols.add("timer-outline");
-        symbols.add("timer-sharp");
-        symbols.add("timer");
-        symbols.add("today-outline");
-        symbols.add("today-sharp");
-        symbols.add("today");
-        symbols.add("toggle-outline");
-        symbols.add("toggle-sharp");
-        symbols.add("toggle");
-        symbols.add("trail-sign-outline");
-        symbols.add("trail-sign-sharp");
-        symbols.add("trail-sign");
-        symbols.add("train-outline");
-        symbols.add("train-sharp");
-        symbols.add("train");
-        symbols.add("transgender-outline");
-        symbols.add("transgender-sharp");
-        symbols.add("transgender");
-        symbols.add("trash-bin-outline");
-        symbols.add("trash-bin-sharp");
-        symbols.add("trash-bin");
-        symbols.add("trash-outline");
-        symbols.add("trash-sharp");
-        symbols.add("trash");
-        symbols.add("trending-down-outline");
-        symbols.add("trending-down-sharp");
-        symbols.add("trending-down");
-        symbols.add("trending-up-outline");
-        symbols.add("trending-up-sharp");
-        symbols.add("trending-up");
-        symbols.add("triangle-outline");
-        symbols.add("triangle-sharp");
-        symbols.add("triangle");
-        symbols.add("trophy-outline");
-        symbols.add("trophy-sharp");
-        symbols.add("trophy");
-        symbols.add("tv-outline");
-        symbols.add("tv-sharp");
-        symbols.add("tv");
-        symbols.add("umbrella-outline");
-        symbols.add("umbrella-sharp");
-        symbols.add("umbrella");
-        symbols.add("unlink-outline");
-        symbols.add("unlink-sharp");
-        symbols.add("unlink");
-        symbols.add("videocam-off-outline");
-        symbols.add("videocam-off-sharp");
-        symbols.add("videocam-off");
-        symbols.add("videocam-outline");
-        symbols.add("videocam-sharp");
-        symbols.add("videocam");
-        symbols.add("volume-high-outline");
-        symbols.add("volume-high-sharp");
-        symbols.add("volume-high");
-        symbols.add("volume-low-outline");
-        symbols.add("volume-low-sharp");
-        symbols.add("volume-low");
-        symbols.add("volume-medium-outline");
-        symbols.add("volume-medium-sharp");
-        symbols.add("volume-medium");
-        symbols.add("volume-mute-outline");
-        symbols.add("volume-mute-sharp");
-        symbols.add("volume-mute");
-        symbols.add("volume-off-outline");
-        symbols.add("volume-off-sharp");
-        symbols.add("volume-off");
-        symbols.add("walk-outline");
-        symbols.add("walk-sharp");
-        symbols.add("walk");
-        symbols.add("wallet-outline");
-        symbols.add("wallet-sharp");
-        symbols.add("wallet");
-        symbols.add("warning-outline");
-        symbols.add("warning-sharp");
-        symbols.add("warning");
-        symbols.add("watch-outline");
-        symbols.add("watch-sharp");
-        symbols.add("watch");
-        symbols.add("water-outline");
-        symbols.add("water-sharp");
-        symbols.add("water");
-        symbols.add("wifi-outline");
-        symbols.add("wifi-sharp");
-        symbols.add("wifi");
-        symbols.add("wine-outline");
-        symbols.add("wine-sharp");
-        symbols.add("wine");
-        symbols.add("woman-outline");
-        symbols.add("woman-sharp");
-        symbols.add("woman");
-
-        Icon icon = AllIcons.Nodes.Plugin;
-        return symbols.stream().map(e -> new Symbol(PREFIX + e + SUFFIX,
-            PREFIX + e, "plugin-ionicons-api", null, icon)).collect(Collectors.toSet());
+        return Ionicons.getAvailableIcons().keySet().stream().map(e -> new Symbol(PREFIX + e + SUFFIX,
+            PREFIX + e, "plugin-ionicons-api", null, readSvg(e))).collect(Collectors.toSet());
     }
 
     private static boolean hasIoniconsApi(Project project) {
@@ -1426,5 +72,27 @@ public class IoniconsApiSymbolFinder implements SymbolFinder {
             e.printStackTrace();
             return false;
         }
+    }
+
+    public static Icon readSvg(String name) {
+        // Get the resource path for the SVG
+        InputStream fileStream = Ionicons.class.getResourceAsStream("/images/symbols/" + name + ".svg");
+
+        if (fileStream == null) {
+            throw new IllegalArgumentException("SVG file not found: " + name);
+        }
+
+        // Use BufferedReader to read the file as a String
+        StringBuilder svgContent = new StringBuilder();
+        try (BufferedReader reader = new BufferedReader(new InputStreamReader(fileStream, StandardCharsets.UTF_8))) {
+            String line;
+            while ((line = reader.readLine()) != null) {
+                svgContent.append(line).append("\n");
+            }
+        } catch (Exception e) {
+            throw new RuntimeException("Error reading SVG file: " + name, e);
+        }
+
+        return convertSymbolToIcon(convertSymbol(svgContent.toString()));
     }
 }
