+++ src/main/java/org/kohsuke/stapler/idea/symbols/LocalSymbolFinder.java
@@ -2,15 +2,12 @@ package org.kohsuke.stapler.idea.symbols;
 
 import com.intellij.openapi.project.Project;
 import com.intellij.openapi.vfs.VirtualFile;
-import com.intellij.util.SVGLoader;
 
-import javax.swing.*;
-import java.io.ByteArrayInputStream;
-import java.nio.charset.StandardCharsets;
 import java.util.HashSet;
 import java.util.Set;
 
 import static org.kohsuke.stapler.idea.ProjectHelper.getArtifactId;
+import static org.kohsuke.stapler.idea.icons.Icons.convertSymbolToIcon;
 
 public class LocalSymbolFinder implements SymbolFinder {
 
@@ -52,40 +49,5 @@ public class LocalSymbolFinder implements SymbolFinder {
         }
 
         return svgFiles;
-    }
-
-    public static Icon convertSymbolToIcon(VirtualFile svgFile) {
-        try {
-            // Ensure the file exists and is not a directory
-            if (svgFile.exists() && !svgFile.isDirectory() && svgFile.getName().endsWith(".svg")) {
-                byte[] contentBytes = svgFile.contentsToByteArray();
-                return convertSymbolToIcon(new String(contentBytes, StandardCharsets.UTF_8)
-                    .replaceAll("var(--blue)", "currentColor")
-                    .replaceAll("var(--yellow)", "currentColor")
-                    .replaceAll("var(--red)", "currentColor")
-                    .replaceAll("var(--text-color-secondary)", "currentColor")
-                    .replaceAll("var(--text-color)", "currentColor")
-                    .replaceAll("currentColor", "#CED0D6"));
-            }
-        } catch (Exception e) {
-            e.printStackTrace();
-        }
-        return null;
-    }
-
-    private static Icon convertSymbolToIcon(String svgContent) {
-        try {
-            // Convert the SVG content into an InputStream
-            ByteArrayInputStream inputStream = new ByteArrayInputStream(svgContent.getBytes(StandardCharsets.UTF_8));
-
-            // Use IntelliJ's SVGLoader to render the icon
-            var image = SVGLoader.load(inputStream, 1f);
-
-            return new ImageIcon(image);
-        } catch (Exception e) {
-            e.printStackTrace();
-        }
-
-        return null;
     }
 }
